test post-cmd changes, fix false alerts on start
In the second pass, after commands have been executed according to the
configuration, the index was not updated to include each files updated
status. This meant that files that were changed, as a result of a
command being ran, would retain their pre-command-execution metadata in
the index. This results in fsautoproc treating that as a modification at
next start, re-running the same processing again each time.